# Frontend/backend_openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Valio Aimo Customer Portal API",
      "description": "Backend API for Zero-Fail Logistics Customer Portal",
      "version": "2.0.0"
    },
    "paths": {
...
      "/customer/payments/invoices": {
        "get": {
          "tags": [
            "customer-payments"
          ],
          "summary": "Get Invoices",
          "description": "Get all active invoices for customer (excludes cancelled invoices)",
          "operationId": "get_invoices_customer_payments_invoices_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/Invoice"
                    },
                    "type": "array",
                    "title": "Response Get Invoices Customer Payments Invoices Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/customer/payments/invoices/{invoice_id}": {
        "get": {
          "tags": [
            "customer-payments"
          ],
          "summary": "Get Invoice",
          "description": "Get specific invoice details",
          "operationId": "get_invoice_customer_payments_invoices__invoice_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "invoice_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Invoice Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "Invoice": {
          "properties": {
            "invoice_id": {
              "type": "string",
              "title": "Invoice Id"
            },
            "order_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            },
            "claim_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claim Id"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            },
            "invoice_type": {
              "$ref": "#/components/schemas/InvoiceType"
            },
            "status": {
              "$ref": "#/components/schemas/InvoiceStatus"
            },
            "total_amount": {
              "type": "number",
              "title": "Total Amount"
            },
            "tax_amount": {
              "type": "number",
              "title": "Tax Amount"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "paid_at": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Paid At"
            },
            "items": {
              "items": {
                "$ref": "#/components/schemas/InvoiceItem"
              },
              "type": "array",
              "title": "Items",
              "default": []
            }
          },
          "type": "object",
          "required": [
            "invoice_id",
            "customer_id",
            "invoice_type",
            "status",
            "total_amount",
            "tax_amount",
            "created_at"
          ],
          "title": "Invoice"
        },
        "InvoiceItem": {
          "properties": {
            "item_id": {
              "type": "integer",
              "title": "Item Id"
            },
            "invoice_id": {
              "type": "string",
              "title": "Invoice Id"
            },
            "product_code": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Code"
            },
            "description": {
              "type": "string",
              "title": "Description"
            },
            "quantity": {
              "type": "number",
              "title": "Quantity"
            },
            "unit_price": {
              "type": "number",
              "title": "Unit Price"
            },
            "total_price": {
              "type": "number",
              "title": "Total Price"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "item_id",
            "invoice_id",
            "description",
            "quantity",
            "unit_price",
            "total_price",
            "created_at"
          ],
          "title": "InvoiceItem"
        },
        "InvoiceStatus": {
          "type": "string",
          "enum": [
            "pending",
            "paid",
            "refunded",
            "cancelled"
          ],
          "title": "InvoiceStatus"
        },
        "InvoiceType": {
          "type": "string",
          "enum": [
            "order",
            "refund",
            "modification"
          ],
          "title": "InvoiceType"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "auth/login"
            }
          }
        }
      }
    }
  }